pyproject.toml
[project]
name = "jcb-use-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=0.28.0",
    "bitsandbytes>=0.48.2",
    "easyocr>=1.7.2",
    "google-genai>=1.55.0",
    "google-generativeai>=0.8.6",
    "gradio>=6.2.0",
    "groq>=0.37.1",
    "huggingface-hub[hf-xet]>=0.23.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "paddleocr>=2.7.0",
    "paddlepaddle>=2.6.0",
    "pdf2image>=1.16.3",
    "peft>=0.18.0",
    "pillow>=10.0.0",
    "pip>=25.3",
    "protobuf>=4.21.0,<6.0.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "regex>=2025.11.3",
    "sentencepiece>=0.2.1",
    "streamlit>=1.52.2",
    "tesseract>=0.1.3",
    "torch>=2.0.0",
    "torchaudio>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers>=4.40.0",
]
